requirements.txt
# Configuração de Ambiente
python-dotenv
requests
pytz
numpy

# LLM / Gemini API
google-genai

# Google Calendar API
google-api-python-client
google-auth
google-auth-oauthlib

# Vector Store
faiss-cpu